Move wzdx_sandbox code to folder
